gene models: refactor of GTF serialization code and unit tests
